Fix join_instances_via_supertype
