+++ mypy/meet.py
@@ -5,7 +5,10 @@ from collections.abc import Callable
 from mypy import join
 from mypy.erasetype import erase_type
 from mypy.maptype import map_instance_to_supertype
-from mypy.disallow_str_iteration_state import disallow_str_iteration_state
+from mypy.disallow_str_iteration_state import (
+    disallow_str_iteration_state,
+    is_subtype_relation_ignored_to_disallow_str_iteration,
+)
 from mypy.state import state
 from mypy.subtypes import (
     are_parameters_compatible,
@@ -15,7 +18,6 @@ from mypy.subtypes import (
     is_proper_subtype,
     is_same_type,
     is_subtype,
-    is_subtype_relation_ignored_to_disallow_str_iteration,
 )
 from mypy.typeops import is_recursive_pair, make_simplified_union, tuple_fallback
 from mypy.types import (
