+++ mypy/subtypes.py
@@ -34,7 +34,10 @@ from mypy.nodes import (
     Var,
 )
 from mypy.options import Options
-from mypy.disallow_str_iteration_state import disallow_str_iteration_state
+from mypy.disallow_str_iteration_state import (
+    disallow_str_iteration_state,
+    is_subtype_relation_ignored_to_disallow_str_iteration,
+)
 from mypy.state import state
 from mypy.types import (
     MYPYC_NATIVE_INT_NAMES,
@@ -1200,6 +1203,11 @@ def is_protocol_implementation(
     as well.
     """
     assert right.type.is_protocol
+    if (
+        disallow_str_iteration_state.disallow_str_iteration
+        and is_subtype_relation_ignored_to_disallow_str_iteration(left, right)
+    ):
+        return False
     if skip is None:
         skip = []
     # We need to record this check to generate protocol fine-grained dependencies.
@@ -2319,21 +2327,3 @@ def is_erased_instance(t: Instance) -> bool:
         elif not isinstance(get_proper_type(arg), AnyType):
             return False
     return True
-
-
-def is_subtype_relation_ignored_to_disallow_str_iteration(left: Instance, right: Instance) -> bool:
-    return (
-        left.type.has_base("builtins.str")
-        and not right.type.has_base("builtins.str")
-        and any(
-            right.type.has_base(base)
-            for base in (
-                "collections.abc.Collection",
-                "collections.abc.Iterable",
-                "collections.abc.Sequence",
-                "typing.Collection",
-                "typing.Iterable",
-                "typing.Sequence",
-            )
-        )
-    )
